metricToImperial() and imperialToMetric() functions intialised in RecipeConversionHelpers class

+++ Backend/src/main/java/com/flavourfit/Recipes/IRecipeService.java
@@ -18,5 +18,5 @@ public interface IRecipeService {
     CompleteRecipeDto recordRecipe(CompleteRecipeDto completeRecipe, int userId) throws
             RecipeExceptions;
 
-    CompleteRecipeDto convertRecipe(int recipeId, int scale, String system) throws RecipeExceptions;
+    CompleteRecipeDto convertRecipe(int recipeId, double scale, String system) throws RecipeExceptions;
 }
